Move from models folder, add Path and HttpOnly props to set-cookie
